backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
openai-whisper==20231117
google-generativeai==0.4.1
langchain==0.1.16
langchain-google-genai==1.0.3
reportlab==4.1.0
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.1
torch==2.2.2
pyannote.audio==3.1.1
faiss-cpu==1.8.0
sentence-transformers==2.6.1
langchain-huggingface==0.0.3
